refactor(fe): unified /authorize route with store-driven rendering
Replace three route groups (channel, pending-channel, resuming-channel)
with a single /authorize route. URL params determine the channel mode,
store state drives which view renders — no child route navigation.
