--- src/frontend/src/routes/(new-styling)/(channel)/authorize/(panel)/+layout.svelte
@@ -1,149 +1,0 @@
-<script lang="ts">
-  import type { LayoutProps } from "./$types";
-  import AuthPanel from "$lib/components/layout/AuthPanel.svelte";
-  import { lastUsedIdentitiesStore } from "$lib/stores/last-used-identities.store";
-  import MigrationIllustration from "$lib/components/illustrations/MigrationIllustration.svelte";
-  import { XIcon } from "@lucide/svelte";
-  import { t } from "$lib/stores/locale.store";
-  import { getDapps } from "$lib/legacy/flows/dappsExplorer/dapps";
-  import { establishedChannelStore } from "$lib/stores/channelStore";
-  import { Trans } from "$lib/components/locale";
-  import { MigrationWizard } from "$lib/components/wizards/migration";
-  import { handleError } from "$lib/components/utils/error";
-  import Dialog from "$lib/components/ui/Dialog.svelte";
-  import { GUIDED_UPGRADE, MIN_GUIDED_UPGRADE } from "$lib/state/featureFlags";
-  import { goto } from "$app/navigation";
-
-  const { children }: LayoutProps = $props();
-
-  let isUpgradeCollapsed = $state(
-    localStorage.getItem("ii-guided-upgrade-collapsed") === "true" ||
-      $MIN_GUIDED_UPGRADE,
-  );
-  let upgradePanelHeight = $state<number>(224);
-  let isUpgrading = $state(false);
-
-  const dapps = getDapps();
-  const dapp = $derived(
-    dapps.find((dapp) => dapp.hasOrigin($establishedChannelStore.origin)),
-  );
-
-  const onUpgrade = async (identityNumber: bigint) => {
-    lastUsedIdentitiesStore.selectIdentity(identityNumber);
-    isUpgrading = false;
-    isUpgradeCollapsed = true;
-    await goto("/authorize/upgrade-success");
-  };
-
-  $effect(() => {
-    localStorage.setItem(
-      "ii-guided-upgrade-collapsed",
-      `${isUpgradeCollapsed}`,
-    );
-  });
-</script>
-
-{#snippet upgradePanel()}
-  <div
-    class={[
-      "relative overflow-hidden rounded-xl border-1 transition-all duration-200 max-sm:mx-4",
-      "border-[#d3d8e6] bg-[#e8e9f2]",
-      "dark:border-[#34384b] dark:bg-[#1f212d]",
-      isUpgradeCollapsed && "sm:-mb-[calc(var(--upgrade-card-height)-22px)]",
-    ]}
-    style="--upgrade-card-height: {upgradePanelHeight}px"
-  >
-    <div
-      bind:clientHeight={upgradePanelHeight}
-      inert={isUpgradeCollapsed}
-      class={[
-        "relative flex flex-col items-stretch p-6 transition-all transition-discrete duration-200",
-        isUpgradeCollapsed
-          ? "opacity-0 max-sm:-mb-[calc(var(--upgrade-card-height)-44px)]"
-          : "z-1 delay-100",
-      ]}
-    >
-      <MigrationIllustration
-        class={[
-          "mb-3 self-center text-black dark:text-white",
-          "!gap-3 !py-1.5 [&_svg:first-child]:!h-3 [&_svg:last-child]:!h-3 [&_svg:nth-child(2)]:!size-3",
-        ]}
-      />
-      <h2
-        class="mx-6 mb-3 text-center text-lg font-medium text-balance text-black dark:text-white"
-      >
-        {#if dapp?.name !== undefined}
-          {@const application = dapp.name}
-          {$t`${application} has moved to the new Internet Identity`}
-        {:else}
-          {$t`This app has moved to the new Internet Identity`}
-        {/if}
-      </h2>
-      <p class="mb-3 text-center text-sm text-black/70 dark:text-white/70">
-        <Trans>Still using an identity number?</Trans>
-      </p>
-      <button
-        onclick={() => (isUpgrading = true)}
-        class="btn !bg-[#6752cc] !text-white hover:!bg-[#6d57cf]"
-      >
-        {$t`Upgrade your identity`}
-      </button>
-      <button
-        onclick={() => (isUpgradeCollapsed = true)}
-        class="btn btn-tertiary btn-lg btn-icon absolute end-2 top-2 !rounded-full hover:!bg-black/5 dark:hover:!bg-white/3"
-        aria-label={$t`Close`}
-      >
-        <XIcon class="size-5" aria-hidden="true" />
-      </button>
-    </div>
-    <div
-      inert={!isUpgradeCollapsed}
-      class={[
-        "absolute inset-x-0 top-0 flex h-11 flex-row items-stretch gap-3 transition-all transition-discrete duration-200",
-        !isUpgradeCollapsed ? "opacity-0" : "z-1 delay-200",
-      ]}
-    >
-      <button
-        onclick={() => (isUpgradeCollapsed = false)}
-        class="flex-1 border-none ps-6 text-start text-sm text-black/70 outline-none dark:text-white/70"
-      >
-        <Trans>Still using an identity number?</Trans>
-      </button>
-      <button
-        onclick={() => (isUpgrading = true)}
-        class="me-6 border-none text-sm font-semibold text-black/70 outline-none hover:underline focus-visible:underline dark:text-white/70"
-      >
-        {$t`Upgrade`}
-      </button>
-    </div>
-  </div>
-  {#if isUpgrading}
-    <Dialog onClose={() => (isUpgrading = false)}>
-      <MigrationWizard
-        onSuccess={onUpgrade}
-        onError={(error) => {
-          handleError(error);
-          isUpgrading = false;
-        }}
-      />
-    </Dialog>
-  {/if}
-{/snippet}
-
-<div class="grid w-full flex-1 items-center max-sm:items-stretch sm:max-w-100">
-  <div class="relative col-start-1 row-start-1 flex flex-col gap-5">
-    {#if $GUIDED_UPGRADE || $MIN_GUIDED_UPGRADE}
-      {@render upgradePanel()}
-    {/if}
-    <AuthPanel
-      class={[
-        "z-1",
-        ($GUIDED_UPGRADE || $MIN_GUIDED_UPGRADE) &&
-          isUpgradeCollapsed &&
-          "rounded-t-none",
-      ]}
-    >
-      {@render children()}
-    </AuthPanel>
-  </div>
-</div>
