--- src/frontend/src/routes/(new-styling)/(channel)/authorize/upgrade-success/+layout.ts
@@ -1,11 +1,0 @@
-import type { LayoutLoad } from "./$types";
-import { redirect } from "@sveltejs/kit";
-import { get } from "svelte/store";
-import { isAuthenticatedStore } from "$lib/stores/authentication.store";
-
-export const load: LayoutLoad = () => {
-  // Go back to /authorize if not authenticated
-  if (!get(isAuthenticatedStore)) {
-    throw redirect(307, "/authorize");
-  }
-};
