--- src/frontend/src/routes/(new-styling)/(pending-channel)/init-openid-authorize/+page.ts
@@ -1,15 +1,0 @@
-import { redirect } from "@sveltejs/kit";
-import { backendCanisterConfig } from "$lib/globals";
-import { PageLoad } from "./$types";
-
-export const load: PageLoad = ({ url }) => {
-  const issuer = url.searchParams.get("openid");
-  const config = backendCanisterConfig.openid_configs[0]?.find(
-    (config) => issuer !== undefined && config.issuer === issuer,
-  );
-  if (config === undefined) {
-    // If OpenID config can't be found, fallback to ICRC-29 authorization flow
-    throw redirect(307, "/authorize");
-  }
-  return { config };
-};
